+++ test/scrivener/paginator/ecto/query_test.exs
@@ -1,7 +1,7 @@
 defmodule Scrivener.Paginater.Ecto.QueryTest do
   use Scrivener.Ecto.TestCase
 
-  alias Scrivener.Ecto.{Comment, KeyValue, Post}
+  alias Scrivener.Ecto.{Comment, KeyValue, Post, User}
 
   defp create_posts do
     unpublished_post = insert_new_post!("Title unpublished", "Body unpublished", false)
@@ -45,6 +45,13 @@ defmodule Scrivener.Paginater.Ecto.QueryTest do
         value: rem(i, 2) |> to_string
       }
       |> Scrivener.Ecto.Repo.insert!()
+    end)
+  end
+
+  defp create_users(number, prefix) do
+    Enum.map(1..number, fn i ->
+      %User{email: "user_#{i}@#{prefix}.com"}
+      |> Scrivener.Ecto.Repo.insert!(prefix: prefix)
     end)
   end
 
@@ -371,6 +378,41 @@ defmodule Scrivener.Paginater.Ecto.QueryTest do
       assert page.total_entries == 6
       assert page.page_size == 5
       assert page.total_pages == 2
+    end
+
+    test "pagination plays nice with absolute distinct on a join query" do
+      create_posts()
+
+      page =
+        Post
+        |> distinct(true)
+        |> join(:inner, [p], c in assoc(p, :comments))
+        |> Scrivener.Ecto.Repo.paginate()
+
+      assert length(page.entries) == 1
+      assert page.page_size == 5
+      assert page.page_number == 1
+      assert page.total_entries == 1
+      assert page.total_pages == 1
+    end
+
+    test "can specify prefix" do
+      create_users(6, "tenant_1")
+      create_users(2, "tenant_2")
+
+      page_tenant_1 = Scrivener.Ecto.Repo.paginate(User, options: [prefix: "tenant_1"])
+
+      assert page_tenant_1.page_size == 5
+      assert page_tenant_1.page_number == 1
+      assert page_tenant_1.total_entries == 6
+      assert page_tenant_1.total_pages == 2
+
+      page_tenant_2 = Scrivener.Ecto.Repo.paginate(User, options: [prefix: "tenant_2"])
+
+      assert page_tenant_2.page_size == 5
+      assert page_tenant_2.page_number == 1
+      assert page_tenant_2.total_entries == 2
+      assert page_tenant_2.total_pages == 1
     end
   end
 end