+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 
-## 2.4.0-dev
+## 2.8.0-dev
+
+## 2.7.0
+
+* Don't pass `prefix` to repo when not provided
+
+## 2.6.0
+
+* Remove elixir 1.11 warnings
+
+## 2.5.0
+
+* Add `prefix` options (see
+  [test](https://github.com/drewolson/scrivener_ecto/blob/8ef17361251aea9784d5b0402547327d90ca4184/test/scrivener/paginator/ecto/query_test.exs#L430)
+  for usage)
+
+## 2.4.0
+
+* Handle absolute distinct clauses in join pagination
 
 ## 2.3.0